backend/requirements.txt
fastapi
uvicorn[standard]
snowflake-id
psycopg[binary]